Move `File` to its own module.
This is a pure refactor in preparation for adding more features to
`File` so that file descriptors can be transferred between processes by
Binder.
