+++ drivers/android/process.rs
@@ -8,7 +8,8 @@ use core::{
 };
 use kernel::{
     bindings, c_types,
-    file_operations::{File, FileOpener, FileOperations, IoctlCommand, IoctlHandler, PollTable},
+    file::File,
+    file_operations::{FileOpener, FileOperations, IoctlCommand, IoctlHandler, PollTable},
     io_buffer::{IoBufferReader, IoBufferWriter},
     linked_list::List,
     pages::Pages,
