+++ rust/kernel/lib.rs
@@ -40,6 +40,7 @@ pub mod buffer;
 pub mod c_types;
 pub mod chrdev;
 mod error;
+pub mod file;
 pub mod file_operations;
 pub mod miscdev;
 pub mod pages;
