+++ samples/rust/rust_semaphore.rs
@@ -23,7 +23,8 @@ use core::{
 };
 use kernel::{
     condvar_init, cstr, declare_file_operations,
-    file_operations::{File, FileOpener, FileOperations, IoctlCommand, IoctlHandler},
+    file::File,
+    file_operations::{FileOpener, FileOperations, IoctlCommand, IoctlHandler},
     io_buffer::{IoBufferReader, IoBufferWriter},
     miscdev::Registration,
     mutex_init,
